Fixed training & API bugs

+++ ml/data.py
@@ -43,7 +43,7 @@ class ReviewDataset(Dataset):
         return {
             'input_ids': encoding['input_ids'].flatten(),
             'attention_mask': encoding['attention_mask'].flatten(),
-            'label': torch.tensor(label, dtype=torch.float)
+            'label': torch.tensor(label, dtype=torch.long) # todo: change uc.
         }
